requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
requests>=2.32.0
python-dotenv>=1.0.0
streamlit>=1.40.0
fastapi>=0.115.0
uvicorn>=0.30.0
